add run modes in config
